Updated button for loop
Added a operator button variable so that the operators can be identified when performing a calculation. It will also recognize if the user has input an operator before a number on the calculation. For example: -1. I also added further IF functionality so that the AC button clears the screen and the = button is not visible.